requirements.txt
alembic==1.18.4
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==26.1.0
Automat==25.4.16
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coloredlogs==14.0
constantly==23.10.4
cryptography==46.0.5
cssselect==1.4.0
dagster==1.12.21
dagster-graphql==1.12.21
dagster-pipes==1.12.21
dagster-shared==1.12.21
dagster-webserver==1.12.21
defusedxml==0.7.1
dnspython==2.8.0
docstring_parser==0.17.0
filelock==3.25.2
fsspec==2026.2.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
grpcio==1.78.0
grpcio-health-checking==1.78.0
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
Jinja2==3.1.6
jmespath==1.1.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minio==7.2.20
multidict==6.7.1
packaging==26.0
parsel==1.11.0
pathlib_abc==0.5.2
platformdirs==4.9.4
propcache==0.4.1
Protego==0.6.0
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyDispatcher==2.0.7
Pygments==2.19.2
pymongo==4.16.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
queuelib==1.9.0
requests==2.33.0
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.3
Scrapy==2.14.2
service-identity==24.2.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
structlog==25.5.0
tabulate==0.10.0
tldextract==5.3.1
tomli==2.4.1
tomlkit==0.14.0
toposort==1.10
tqdm==4.67.3
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
universal_pathlib==0.3.10
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
w3lib==2.4.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
zope.interface==8.2
